+++ scripts/release/README.md
@@ -4,14 +4,14 @@ The release process consists of several phases, each one represented by one of t
 
 A typical release goes like this:
 1. When a commit is pushed to the React repo, [Circle CI](https://circleci.com/gh/facebook/react/) will build all release bundles and run unit tests against both the source code and the built bundles.
-2. The release is then [**published to the `next` channel**](#publishing-release) using the [`download-release`](#download-release) and [`publish`](#publish) scripts. (Currently this process is manual but might be automated in the future using [GitHub "actions"](https://github.com/features/actions).)
+2. The release is then [**published to the `next` channel**](#publishing-release) using the [`prepare-release-from-ci`](#prepare-release-from-ci) and [`publish`](#publish) scripts. (Currently this process is manual but might be automated in the future using [GitHub "actions"](https://github.com/features/actions).)
    1. The release may also be [**published to the `experimental` channel**](#publishing-an-experimental-release) using the the same scripts (but different build artifacts).
-3. Finally, a "next" release can be [**promoted to stable**](#publishing-a-stable-release)<sup>1</sup> using the [`promote-to-stable`](#promote-to-stable) and [`publish`](#publish) scripts. (This process is always manual.)
+3. Finally, a "next" release can be [**promoted to stable**](#publishing-a-stable-release)<sup>1</sup> using the [`prepare-release-from-npm`](#prepare-release-from-npm) and [`publish`](#publish) scripts. (This process is always manual.)
 
 The high level process of creating releases is [documented below](#process). Individual scripts are documented as well:
 * [`build-release-locally`](#build-release-locally): Build a release locally from the checked out source code.
-* [`download-release`](#download-release): Download a pre-built release from CI.
-* [`promote-to-stable`](#promote-to-stable): Prepare an NPM "next" release to be published as a "stable" release.
+* [`prepare-release-from-ci`](#prepare-release-from-ci): Download a pre-built release from CI.
+* [`prepare-release-from-npm`](#prepare-release-from-npm): Prepare an NPM "next" release to be published as a "stable" release.
 * [`publish`](#publish): Publish the downloaded (or prepared) release to NPM.
 
 <sup>1. [**Creating a patch release**](#creating-a-patch-release) has a slightly different process than a major/minor release.</sup>
@@ -31,9 +31,9 @@ To prepare a build for a particular commit:
    * If it's still pending, you'll need to wait for it to finish. <sup>1</sup>
 4. Copy the build ID from the URL
    * e.g. the build ID for commit [e5d06e34b](https://github.com/facebook/react/commit/e5d06e34b) is [**124756**](https://circleci.com/gh/facebook/react/124756)
-5. Run the [`download-release`](#download-release) script with the build ID <sup>2</sup> you found:
+5. Run the [`prepare-release-from-ci`](#prepare-release-from-ci) script with the build ID <sup>2</sup> you found:
 ```sh
-scripts/release/download-release.js --build=124756
+scripts/release/prepare-release-from-ci.js --build=124756
 ```
 
 Once the build has been checked out and tested locally, you're ready to publish it:
@@ -54,7 +54,7 @@ The steps for publishing an experimental release are almost the same as for publ
 
 For example, the experimental build ID for commit [e5d06e34b](https://github.com/facebook/react/commit/e5d06e34b) is [**124763**](https://circleci.com/gh/facebook/react/124763):
 ```sh
-scripts/release/download-release.js --build=124763
+scripts/release/prepare-release-from-ci.js --build=124763
 ```
 
 Once the build has been checked out and tested locally, you're ready to publish it. When publishing an experimental release, use the `experimental` tag:
@@ -71,10 +71,10 @@ If the OTP code expires while publishing, re-run this command and answer "y" to 
 
 Stable releases should always be created from the "next" channel. This encourages better testing of the actual release artifacts and reduces the chance of unintended changes accidentally being included in a stable release.
 
-To prepare a stable release, choose a "next" version and run the [`promote-to-stable`](#promote-to-stable) script <sup>1</sup>:
+To prepare a stable release, choose a "next" version and run the [`prepare-release-from-npm`](#prepare-release-from-npm) script <sup>1</sup>:
 
 ```sh
-scripts/release/promote-to-stable.js --version=0.0.0-241c4467e
+scripts/release/prepare-release-from-npm.js --version=0.0.0-241c4467e
 ```
 
 This script will prompt you to select stable version numbers for each of the packages. It will update the package JSON versions (and dependencies) based on the numbers you select.
@@ -132,7 +132,7 @@ To create a build from the current branch and revision:
 scripts/release/build-release-locally.js
 ```
 
-## `download-release`
+## `prepare-release-from-ci`
 Downloads build artifacts from Circle CI in preparation to be published to NPM as either a "next" or "experimental" release.
 
 All artifacts built by Circle CI have already been unit-tested (both source and bundles) but these candidates should **always be manually tested** before being published. Upon completion, this script prints manual testing instructions.
@@ -140,10 +140,10 @@ All artifacts built by Circle CI have already been unit-tested (both source and 
 #### Example usage
 To prepare the artifacts created by [Circle CI build 124756](https://circleci.com/gh/facebook/react/124756#artifacts/containers/0) you would run:
 ```sh
-scripts/release/download-release.js --build=124756
+scripts/release/prepare-release-from-ci.js --build=124756
 ```
 
-## `promote-to-stable`
+## `prepare-release-from-npm`
 Checks out a "next" release from NPM and prepares it to be published as a stable release.
 
 This script prompts for new (stable) release versions for each public package and updates the package contents (both `package.json` and inline version numbers) to match. It also updates inter-package dependencies to account for the new versions.
@@ -153,7 +153,7 @@ This script prompts for new (stable) release versions for each public package an
 #### Example usage
 To promote the "next" release `0.0.0-241c4467e` (aka commit [241c4467e](https://github.com/facebook/react/commit/241c4467e)) to stable:
 ```sh
-scripts/release/promote-to-stable.js --version=0.0.0-241c4467e
+scripts/release/prepare-release-from-npm.js --version=0.0.0-241c4467e
 ```
 
 ## `publish`
