+++ scripts/release/prepare-release-from-npm-commands/confirm-stable-version-numbers.js
