+++ scripts/release/prepare-release-from-npm-commands/get-latest-next-version.js
