+++ scripts/release/prepare-release-from-npm-commands/guess-stable-version-numbers.js
