+++ scripts/release/prepare-release-from-npm-commands/update-stable-version-numbers.js
